+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Yii Framework 2 swiftmailer extension Change Log
 -----------------------
 
 - Enh #37: `yii\swiftmailer\Logger` now chooses logging level depending on incoming entry format (klimov-paul)
+- Enh #40: Added `yii\swiftmailer\Message::setHeaders()` allowing to setup custom headers in batch (klimov-paul)
 
 
 2.0.6 September 09, 2016